+++ plugins/postcss-custom-properties/src/lib/get-custom-properties-from-root.ts
@@ -19,7 +19,7 @@ export default function getCustomPropertiesFromRoot(root, opts): Map<string, val
 		// for each custom property
 		if (customPropertiesObject) {
 			rule.nodes.slice().forEach(decl => {
-				if (isCustomDecl(decl) && !isBlockIgnored(decl)) {
+				if (decl.variable && !isBlockIgnored(decl)) {
 					const { prop } = decl;
 
 					// write the parsed value to the custom property
@@ -54,14 +54,10 @@ export default function getCustomPropertiesFromRoot(root, opts): Map<string, val
 // match html and :root rules
 const htmlSelectorRegExp = /^html$/i;
 const rootSelectorRegExp = /^:root$/i;
-const customPropertyRegExp = /^--[A-z][\w-]*$/;
 
 // whether the node is an html or :root rule
 const isHtmlRule = node => node.type === 'rule' && node.selector.split(',').some(item => htmlSelectorRegExp.test(item)) && Object(node.nodes).length;
 const isRootRule = node => node.type === 'rule' && node.selector.split(',').some(item => rootSelectorRegExp.test(item)) && Object(node.nodes).length;
-
-// whether the node is an custom property
-const isCustomDecl = node => node.type === 'decl' && customPropertyRegExp.test(node.prop);
 
 // whether the node is a parent without children
 const isEmptyParent = node => Object(node.nodes).length === 0;